DF-11834 DF-13319 DF-13320 Fix reflection for some enum fields in UIControlBackground
Former-commit-id: 175810f9a6929386258bba7b1f1dc09959312d7e

+++ Sources/Internal/UI/Styles/UIStyleSheetPropertyDataBase.cpp
@@ -32,7 +32,7 @@ UIStyleSheetPropertyDataBase::UIStyleSheetPropertyDataBase()
                      UIStyleSheetPropertyDescriptor(&controlGroup, "noInput", false),
                      UIStyleSheetPropertyDescriptor(&controlGroup, "exclusiveInput", false),
 
-                     UIStyleSheetPropertyDescriptor(&bgGroup, "drawType", static_cast<int32>(UIControlBackground::DRAW_ALIGNED)),
+                     UIStyleSheetPropertyDescriptor(&bgGroup, "drawType", UIControlBackground::DRAW_ALIGNED),
                      UIStyleSheetPropertyDescriptor(&bgGroup, "sprite", FilePath()),
                      UIStyleSheetPropertyDescriptor(&bgGroup, "frame", 0),
                      UIStyleSheetPropertyDescriptor(&bgGroup, "mask", FilePath()),
